+++ rage/src/bin/rage-keygen/main.rs
@@ -45,6 +45,7 @@ fn main() {
 
     let requested_languages = DesktopLanguageRequester::requested_languages();
     i18n_embed::select(&*LANGUAGE_LOADER, &TRANSLATIONS, &requested_languages).unwrap();
+    age::localizer().select(&requested_languages).unwrap();
 
     let opts = AgeOptions::parse_args_default_or_exit();